requirements.txt
google-cloud-speech
google-cloud-texttospeech
moviepy 
openai
streamlit
pydub 
python-dotenv